refactor: move generateRequest helpers to api-reference
