--- packages/api-client/src/hooks/index.ts
@@ -1,2 +1,0 @@
-// export { useCopilot } from './useCopilot'
-export { useOperation, type ParamMap } from './useOperation'
