+++ packages/api-client/src/index.ts
@@ -1,7 +1,6 @@
 export { ApiClient } from './components/ApiClient'
 
 export * from './helpers'
-export * from './hooks'
 export * from './stores/apiClientStore'
 export * from './stores/apiClientRequestStore'
 export * from './types'
