+++ packages/api-reference/src/helpers/generateRequest.test.ts
@@ -1,0 +1,10 @@
+import { generateRequest } from 'src/helpers/generateRequest'
+import { describe, expect, it } from 'vitest'
+
+describe('generateRequest', () => {
+  it('sets example values', () => {
+    expect(generateRequest({}, {}, 'https://example.com')).toMatchObject({
+      id: 10,
+    })
+  })
+})
