+++ packages/api-reference/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useDarkModeState'
+export * from './useOperation'
 export * from './useRefOnMount'
